Added cleaner code documentation.
Changed the model from zero-inflated to lm (zero inflated was taking forever).

Updated the ggplot to have colors based on classification.

Changed the model from being city based to arrival / departure delay (to ensure there was enough data).

Fixed the call to pivot table, and added a progress indicator that still is a little funky.